cdc: move old value code to a separate file
