+++ components/cdc/src/lib.rs
@@ -9,6 +9,7 @@ mod endpoint;
 mod errors;
 pub mod metrics;
 mod observer;
+mod old_value;
 mod service;
 
 pub use endpoint::{CdcTxnExtraScheduler, Endpoint, Task, Validate};
